+++ src/fallback.rs
@@ -1,10 +1,7 @@
 // TODO: Try boxing the closure to see if we can hide the type
 // TODO: Or maybe use a function pointer?
 
-pub struct Bytes<F>
-where
-    F: Fn(u8) -> bool,
-{
+pub struct Bytes<F> {
     fallback: F,
 }
 
@@ -12,12 +9,12 @@ impl<F> Bytes<F>
 where
     F: Fn(u8) -> bool,
 {
-    pub /* const */ fn new(fallback: F) -> Self {
+    pub const fn new(fallback: F) -> Self {
         Bytes { fallback }
     }
 
     pub fn find(&self, haystack: &[u8]) -> Option<usize> {
-        haystack.iter().cloned().position(&self.fallback)
+        haystack.iter().copied().position(&self.fallback)
     }
 }
 
@@ -26,12 +23,12 @@ pub struct ByteSubstring<'a> {
 }
 
 impl<'a> ByteSubstring<'a> {
-    pub /* const */ fn new(needle: &'a[u8]) -> Self {
+    pub const fn new(needle: &'a[u8]) -> Self {
         ByteSubstring { needle }
     }
 
     #[cfg(feature = "pattern")]
-    pub fn needle_len(&self) -> usize {
+    pub const fn needle_len(&self) -> usize {
         self.needle.len()
     }